refactor: menu의 등록과 상세보기(수정) 페이지의 컴포넌트 공통 사용으로 page 병합
