+++ src/pages/menu/components/index.ts
@@ -1,3 +1,4 @@
 export { default as MenuOption } from "./menuOption";
 export { default as LabelInput } from "./labelInput";
 export { default as LabelTextarea } from "./labelTextarea";
+export { default as OptionModal } from "./optionModal";
