+++ src/pages/menu/components/labelInput/LabelInput.types.ts
