+++ src/pages/menu/detail.tsx
